Renamed a namespace.
