+++ ISAAR.MSolve.Solvers/DomainDecomposition/Dual/Pcg/IInterfaceProjection.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using ISAAR.MSolve.LinearAlgebra.Vectors;
 
-namespace ISAAR.MSolve.Solvers.DomainDecomposition.Dual.Pcpg
+namespace ISAAR.MSolve.Solvers.DomainDecomposition.Dual.Pcg
 {
     internal interface IInterfaceProjection
     {
